+++ paddlemetrics/nominal/theils_u.py
@@ -81,7 +81,7 @@ class TheilsU(Metric):
     def update(self, preds: paddle.Tensor, target: paddle.Tensor) -> None:
         """Update state with predictions and targets."""
         confmat = _theils_u_update(preds, target, self.num_classes, self.nan_strategy, self.nan_replace_value)
-        self.confmat += confmat
+        self.confmat += confmat.cast(self.confmat.dtype)
 
     def compute(self) -> paddle.Tensor:
         """Compute Theil's U statistic."""
